+++ test/src/reflection/user_with_reflection.g.dart
@@ -1,11 +1,12 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/2.1.6
+// BUILDER: reflection_factory/2.2.0
 // BUILD COMMAND: dart run build_runner build
 //
 
 // coverage:ignore-file
 // ignore_for_file: unused_element
+// ignore_for_file: no_leading_underscores_for_local_identifiers
 // ignore_for_file: unnecessary_const
 // ignore_for_file: unnecessary_cast
 // ignore_for_file: unnecessary_type_check
@@ -17,7 +18,7 @@ typedef __TI<T> = TypeInfo<T>;
 typedef __PR = ParameterReflection;
 
 mixin __ReflectionMixin {
-  static final Version _version = Version.parse('2.1.6');
+  static final Version _version = Version.parse('2.2.0');
 
   Version get reflectionFactoryVersion => _version;
 
@@ -146,7 +147,7 @@ class TestAddressWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestAddressWithReflection$reflection withObject(
@@ -542,7 +543,7 @@ class TestCompanyWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestCompanyWithReflection$reflection withObject(
@@ -1001,7 +1002,7 @@ class TestDataWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestDataWithReflection$reflection withObject([TestDataWithReflection? obj]) =>
@@ -1299,7 +1300,7 @@ class TestDomainWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestDomainWithReflection$reflection withObject(
@@ -1748,7 +1749,7 @@ class TestEmpty$reflection extends ClassReflection<TestEmpty>
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestEmpty$reflection withObject([TestEmpty? obj]) =>
@@ -1890,7 +1891,7 @@ class TestEnumWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestEnumWithReflection$reflection withObject([TestEnumWithReflection? obj]) =>
@@ -1977,7 +1978,7 @@ class TestFranchiseWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestFranchiseWithReflection$reflection withObject(
@@ -2323,7 +2324,7 @@ class TestName$reflection extends ClassReflection<TestName>
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestName$reflection withObject([TestName? obj]) =>
@@ -2625,7 +2626,7 @@ class TestOpAWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestOpAWithReflection$reflection withObject([TestOpAWithReflection? obj]) =>
@@ -2989,7 +2990,7 @@ class TestOpBWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestOpBWithReflection$reflection withObject([TestOpBWithReflection? obj]) =>
@@ -3371,7 +3372,7 @@ class TestOpWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestOpWithReflection$reflection withObject([TestOpWithReflection? obj]) =>
@@ -3772,7 +3773,7 @@ class TestTransactionWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestTransactionWithReflection$reflection withObject(
@@ -4044,7 +4045,7 @@ class TestUserWithReflection$reflection
   }
 
   @override
-  Version get languageVersion => Version.parse('2.19.0');
+  Version get languageVersion => Version.parse('3.0.0');
 
   @override
   TestUserWithReflection$reflection withObject([TestUserWithReflection? obj]) =>